Update document.title with timer value.
